user essentials mob! :trident:
lastFile:public/app/core/components/AppChrome/DockedMegaMenu/DockedMegaMenu.tsx
